Call allocation listener on BaseAllocator#wrapForeignAllocation
