+++ java/memory/memory-core/src/main/java/org/apache/arrow/memory/BufferAllocator.java
@@ -248,22 +248,5 @@ public interface BufferAllocator extends AutoCloseable {
    *
    * @param allocation The underlying allocation.
    */
-  default ArrowBuf wrapForeignAllocation(ForeignAllocation allocation) {
-    try {
-      forceAllocate(allocation.getSize());
-      final AllocationManager manager = new ForeignAllocationManager(this, allocation);
-      final BufferLedger ledger = manager.associate(this);
-      final ArrowBuf buf =
-          new ArrowBuf(ledger, /*bufferManager=*/null, allocation.getSize(), allocation.memoryAddress());
-      buf.writerIndex(allocation.getSize());
-      return buf;
-    } catch (Throwable t) {
-      try {
-        allocation.release0();
-      } catch (Throwable e) {
-        t.addSuppressed(e);
-      }
-      throw t;
-    }
-  }
+  ArrowBuf wrapForeignAllocation(ForeignAllocation allocation);
 }
